renamed store to useRequestsStore
