+++ src/app/dashboard/requests/RequestsClient.tsx
@@ -9,7 +9,7 @@ import type { DataSectionProps } from "./DataSection";
 import { Requests } from "./Requests";
 
 import { useAuth } from "@/lib/auth";
-import { useRequestListStore } from "@/lib/stores/useRequestListStore";
+import { useRequestListStore } from "@/lib/stores/useRequestStore";
 
 export interface RequestItem {
   id: string;
